# todo/todo-service/swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  TodoMedia:
    description: TodoMedia media type (default view)
    example:
      completedAt: 1.7849052587171064e+18
      createdAt: 5.1631909098674925e+17
      description: Id dolores ut iusto sit.
      done: true
      id: Sint ipsa.
      title: Et cumque eum ut suscipit.
    properties:
      completedAt:
        example: 1.7849052587171064e+18
        format: int64
        type: integer
      createdAt:
        example: 5.1631909098674925e+17
        format: int64
        type: integer
      description:
        example: Id dolores ut iusto sit.
        type: string
      done:
        example: true
        type: boolean
      id:
        example: Sint ipsa.
        type: string
      title:
        example: Et cumque eum ut suscipit.
        type: string
    required:
    - id
    - createdAt
    - done
    title: 'Mediatype identifier: application/json; view=default'
    type: object
  TodoPayload:
    description: Todo payload
    example:
      description: Explicabo et.
      title: Enim placeat et culpa sed minus.
    properties:
      description:
        description: Todo description
        example: Explicabo et.
        type: string
      title:
        description: Todo title
        example: Enim placeat et culpa sed minus.
        type: string
    required:
    - title
    - description
    title: TodoPayload
    type: object
  TodoUpdatePayload:
    description: Todo update payload
    example:
      description: Et reprehenderit officia aut.
      done: true
      title: Deserunt ad placeat.
    properties:
      description:
        description: Todo description
        example: Et reprehenderit officia aut.
        type: string
      done:
        description: Todo status
        example: true
        type: boolean
      title:
        description: Todo title
        example: Deserunt ad placeat.
        type: string
    title: TodoUpdatePayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609066e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609066e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
host: localhost:8080
info:
  title: Micro Todo
  version: ""
paths:
  /todo/{todoID}:
    get:
      description: Get todo by ID
      operationId: todo#getById
      parameters:
      - description: Todo ID
        in: path
        name: todoID
        required: true
        type: string
      produces:
      - application/json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TodoMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: getById todo
      tags:
      - todo
    patch:
      description: Update todo
      operationId: todo#updateTodo#1
      parameters:
      - description: Todo ID
        in: path
        name: todoID
        required: true
        type: string
      - description: Todo update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TodoUpdatePayload'
      produces:
      - application/json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TodoMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: updateTodo todo
      tags:
      - todo
    put:
      description: Update todo
      operationId: todo#updateTodo
      parameters:
      - description: Todo ID
        in: path
        name: todoID
        required: true
        type: string
      - description: Todo update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TodoUpdatePayload'
      produces:
      - application/json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TodoMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: updateTodo todo
      tags:
      - todo
  /todo/{todoID}/delete:
    delete:
      description: Delete todo
      operationId: todo#deleteTodo
      parameters:
      - description: Todo ID
        in: path
        name: todoID
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: deleteTodo todo
      tags:
      - todo
  /todo/add:
    post:
      description: Add new todo
      operationId: todo#addTodo
      parameters:
      - description: Todo payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TodoPayload'
      produces:
      - application/json
      - application/vnd.goa.error
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/TodoMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: addTodo todo
      tags:
      - todo
  /todo/all:
    get:
      description: Get all todos
      operationId: todo#getAllTodos
      parameters:
      - description: Limit todos per page
        in: query
        name: limit
        required: false
        type: integer
      - description: number of todos to skip
        in: query
        name: offset
        required: false
        type: integer
      - description: order by
        in: query
        name: order
        required: false
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sorting
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: getAllTodos todo
      tags:
      - todo
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  OK:
    description: OK
schemes:
- http
swagger: "2.0"